union all bots
